tests/test_summarizer.py
"""Verification test for GroundedSummarizer."""
import sys
import json
from pathlib import Path

PROJECT_ROOT = Path(__file__).resolve().parent.parent
sys.path.insert(0, str(PROJECT_ROOT / "src"))

from dotenv import load_dotenv
load_dotenv(PROJECT_ROOT / ".env")

from summarizer import GroundedSummarizer, SummaryResponse, SUMMARIZATION_SYSTEM_PROMPT
from retriever import BM25Retriever, HybridRetriever, load_chunks

CHUNKS_PATH = str(PROJECT_ROOT / "outputs" / "chunks_semantic.json")
chunks = load_chunks(CHUNKS_PATH)

# =========================================================================
# TEST 1: SummaryResponse dataclass
# =========================================================================
print("=" * 72)
print("  TEST 1: SummaryResponse Dataclass")
print("=" * 72)

sr = SummaryResponse(topic="test", overview="A test.", bullets=["bullet [id1]"])
assert sr.topic == "test"
assert sr.is_partial == False
assert sr.refused == False
assert sr.missing_topics == []
d = sr.to_dict()
assert isinstance(d, dict)
assert "overview" in d and "bullets" in d and "chunk_ids" in d
print(f"  [OK] Dataclass fields: {sorted(d.keys())}")

# =========================================================================
# TEST 2: System prompt design
# =========================================================================
print("\n" + "=" * 72)
print("  TEST 2: System Prompt Validation")
print("=" * 72)

prompt = SUMMARIZATION_SYSTEM_PROMPT
assert "ONLY from the context" in prompt, "Must enforce grounding"
assert "chunk_id" in prompt, "Must require chunk_id citations"
assert "MISSING" in prompt, "Must require missing identification"
assert "could not find" in prompt.lower(), "Must preserve refusal phrase"
assert "JSON" in prompt, "Must enforce JSON output"
print("  [OK] Grounding enforced")
print("  [OK] Citation requirement present")
print("  [OK] Missing section required")
print("  [OK] Refusal phrase preserved")
print("  [OK] JSON format enforced")

# =========================================================================
# TEST 3: Live summarization (in-scope)
# =========================================================================
print("\n" + "=" * 72)
print("  TEST 3: Live Summarization (In-Scope)")
print("=" * 72)

hybrid = HybridRetriever(CHUNKS_PATH)
try:
    summarizer = GroundedSummarizer(retriever=hybrid, top_k=8)
    result = summarizer.summarize("cell membrane")

    print(f"  Topic:      {result.topic}")
    print(f"  Refused:    {result.refused}")
    print(f"  Partial:    {result.is_partial}")
    print(f"  Chunks:     {result.n_chunks_used}")
    print(f"  Citations:  {result.chunk_ids}")
    print(f"  Missing:    {result.missing_topics}")
    print(f"  Overview:   {result.overview[:120]}...")
    print(f"  Bullets:    {len(result.bullets)}")
    for b in result.bullets:
        print(f"    - {b[:100]}...")

    assert not result.refused, "In-scope topic should not be refused"
    assert len(result.overview) > 20, "Overview should have content"
    assert len(result.bullets) >= 1, "Should have at least 1 bullet"
    assert len(result.chunk_ids) >= 1, "Should have at least 1 citation"
    print("  [OK] In-scope summary generated with citations")

except Exception as e:
    print(f"  [SKIP] API error (expected if quota exhausted): {e}")

# =========================================================================
# TEST 4: Out-of-scope topic (fast refusal)
# =========================================================================
print("\n" + "=" * 72)
print("  TEST 4: Out-of-Scope Topic")
print("=" * 72)

try:
    result_oos = summarizer.summarize("quantum entanglement")
    print(f"  Refused:  {result_oos.refused}")
    print(f"  Partial:  {result_oos.is_partial}")
    print(f"  Overview: {result_oos.overview}")

    # Cross-encoder scores for OOS should be ~0.0, below threshold
    assert result_oos.refused or result_oos.is_partial, "OOS should refuse or be partial"
    print("  [OK] Out-of-scope topic handled correctly")

except Exception as e:
    print(f"  [SKIP] API error: {e}")

# =========================================================================
# TEST 5: to_dict() serialization
# =========================================================================
print("\n" + "=" * 72)
print("  TEST 5: JSON Serialization")
print("=" * 72)

sr = SummaryResponse(
    topic="test",
    overview="The cell membrane is a semi-permeable barrier [sem_0023].",
    bullets=["Controls entry/exit of substances [sem_0023]"],
    chunk_ids=["sem_0023"],
    is_partial=False,
    model="test",
    n_chunks_used=5,
)
serialized = json.dumps(sr.to_dict(), indent=2)
assert json.loads(serialized)  # valid JSON
print(f"  [OK] Serializable to JSON ({len(serialized)} chars)")

# =========================================================================
print("\n" + "=" * 72)
print("  ALL TESTS PASSED [OK]")
print("=" * 72)
